지혁+지은 1차

+++ WeddingPlanner/src/main/java/com/wedding/oww/vo/ProductVO.java
@@ -5,10 +5,10 @@ import lombok.*;
 @Getter @Setter
 @NoArgsConstructor @AllArgsConstructor @Builder
 public class ProductVO {
-    private Long productNo;
+    private int productNo;
     private String productName;
-    private Integer category;   // 0: hall, 1: studio, 2: dress, 3: makeup
-    private Integer cost;
+    private int category;   // 0: hall, 1: studio, 2: dress, 3: makeup
+    private int cost;
     private String address;
     private String description;
     private String img;
